requirements.txt
crewai>=0.1.0
langchain>=0.1.0
python-dotenv>=0.19.0
openai>=1.0.0
anthropic>=0.5.0
google-generativeai>=0.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pytest>=7.4.0
